Module umm::constants

A module defining a bunch of constant values to be used throughout

Structs

BUILD_DIR

Directory to store compiler artifacts

JAVA_TS_LANG

Reference to treesitter language struct

LIB_DIR

Directory for libraries, jars

POSTGREST_CLIENT

PostGrest client

RETRIEVAL_MESSAGE_INTRO

Retrieval System Message intro

RETRIEVAL_MESSAGE_OUTRO

Retrieval System Message outro

ROOT_DIR

Path to project root

RUNTIME

Runtime

SCRIPT_AST

Rhai script as a AST, behind an mutex.

SEPARATOR

Platform specific separator character for javac paths

SOURCE_DIR

Directory for source files

SUPABASE_KEY

Supabase public api key

SYSTEM_MESSAGE

Entire ChatGPT System Message

SYSTEM_MESSAGE_INTRO

ChatGPT System Message intro

SYSTEM_MESSAGE_OUTRO

ChatGPT System Message outro

TEST_DIR

Directory for test files

UMM_DIR

Directory for umm artifacts

Constants

CLASSNAME_QUERY

Tree-sitter query that returns name of the class

CLASS_CONSTRUCTOR_QUERY

Tree-sitter query that returns class constructor signatures

CLASS_DECLARATION_QUERY

Tree-sitter query that returns class declaration statements

CLASS_FIELDS_QUERY

field: entire field declaration

CLASS_METHOD_QUERY

Tree-sitter query that returns class method signatures

COURSE

Current course. TODO: Move this to init script

IMPORT_QUERY

Tree-sitter query that returns imports made

INTERFACENAME_QUERY

Tree-sitter query that returns name of the interface

INTERFACE_CONSTANTS_QUERY

Tree-sitter query that returns interface constants

INTERFACE_DECLARATION_QUERY

Tree-sitter query that returns interface declaration statements

INTERFACE_METHODS_QUERY

Tree-sitter query that returns interface methods signatures

JUNIT_PLATFORM

file name for JUnit platform console standard jar

MAIN_METHOD_QUERY

Tree-sitter query to check the existence of a main method.

METHOD_CALL_QUERY

Tree-sitter query that returns method call identifiers

PACKAGE_QUERY

Tree-sitter query that returns name of the package

PROMPT_TRUNCATE

Prompt truncation length

TERM

Current term. TODO: Move this to init script

TEST_ANNOTATION_QUERY

Tree-sitter query that returns name of the JUnit @Test annotated methods

Statics

USE_ACTIVE_RETRIEVAL

Whether to use active retrieval or heuristic based retrieval