Function umm::grade::show_result

pub fn show_result(results: Array, gradescope_config: Map) -> Result<()>

Print grade result

results: array of GradeResults to print in a table.

gradescope_config: map of gradescope configuration options, which can contain:

source_files: array of source files to provide feedback on in the submission. Defaults to empty array.

test_files: array of test files to provide feedback on in the submission. Defaults to empty array.

project_title: title of the project. Defaults to empty string.

project_description: description of the project. Defaults to empty string.

pass_threshold: threshold for passing the project. Defaults to 0.7.

show_table: whether to show the grading table. Defaults to true.

results_json: whether to write the gradescope results in JSON format. Defaults to false.

feedback: whether to provide feedback on penalties to students. Defaults to false.

leaderboard: whether to produce leaderboard entries. Also produces relevant SLO feedback. Defaults to false.

debug: whether to write gradescope JSON within the current directory. Defaults to false.

slo_algorithmic_solutions: whether to provide feedback on Algorithmic Solutions SLO. Defaults to false.

slo_code_readability: whether to provide feedback on Code Readability and Formatting SLO. Defaults to false.

slo_comments: whether to provide feedback on Comments SLO. Defaults to false.

slo_error_handling: whether to provide feedback on Error Handling SLO. Defaults to false.

slo_logic: whether to provide feedback on Logic SLO. Defaults to false.

slo_naming_conventions: whether to provide feedback on Naming Conventions SLO. Defaults to false.

slo_oop_programming: whether to provide feedback on Object Oriented Programming SLO. Defaults to false.

slo_syntax: whether to provide feedback on Syntax SLO. Defaults to false.

slo_testing: whether to provide feedback on Testing SLO. Defaults to false.