Struct umm::grade::Query

pub struct Query {
    query: String,
    capture: String,
    filter: Option<FnPtr>,
}

A struct to represet a treesitter query.

Fields
query: String

The query to run.

capture: String

The capture to extract from the query.

filter: Option<FnPtr>

A function pointer to filter the matches using. Must return a boolean.

Implementations

impl Query

pub fn new() -> Self

Creates a new query with default values (empty strings).

pub fn query(&self) -> String

Gets the query to run.

pub fn set_query(self, query: String) -> Self

Sets the query to run.

pub fn capture(&self) -> String

Gets the captures to extract from the query.

pub fn set_capture(self, capture: String) -> Self

Sets the captures to extract from the query.

pub fn filter(&self) -> Option<FnPtr>

Gets the function to filter the results of the query.

pub fn set_filter(self, filter: FnPtr) -> Self

Set the function to filter the results of the query.