Struct umm::grade::QueryGrader

pub struct QueryGrader {
    req_name: String,
    out_of: f64,
    queries: Vec<Query>,
    project: Project,
    file: String,
    constraint: QueryConstraint,
    reason: String,
}

A struct to represent a query grader.

Fields
req_name: String

The name of the requirement.

out_of: f64

The grade for the requirement.

queries: Vec<Query>

The queries to run.

project: Project

The input to run the queries on.

file: String

The file to run the query on.

constraint: QueryConstraint

The constraint of the query.

reason: String

The reason to share with the student.

Implementations

impl QueryGrader

pub fn new() -> Self

Creates a new query grader with default values.

pub fn req_name(&self) -> &str

Gets the name of the requirement.

pub fn set_req_name(self, req_name: String) -> Self

Sets the name of the requirement.

pub fn out_of(&self) -> f64

Gets the “out of” grade for the requirement.

pub fn set_out_of(self, out_of: f64) -> Self

Sets the “out of” grade for the requirement.

pub fn file(&self) -> &str

Gets the file to run the query on.

pub fn set_file(self, file: String) -> Self

Sets the file to run the query on.

pub fn project(&self) -> &Project

Gets the project to run the query on.

pub fn set_project(self, project: Project) -> Self

Sets the project to run the query on.

pub fn queries(&self) -> Vec<Query>

Gets the queries to run.

pub fn constraint(&self) -> QueryConstraint

Gets the constraint of the query.

pub fn must_match_at_least_once(self) -> Self

Sets the constraint of the query to “must match at least once”.

pub fn must_match_exactly_n_times(self, n: usize) -> Self

Sets the constraint of the query to “must match exactly n times”.

pub fn must_not_match(self) -> Self

Sets the constraint of the query to “must not match”.

pub fn reason(&self) -> &str

Gets the reason to share with the student.

pub fn set_reason(self, reason: String) -> Self

Sets the reason to share with the student.

pub fn query(self, q: String) -> Result<Self, QueryError>

Adds a query to run. If no file has been selected, this will throw an error.

pub fn query_script(self, q: String) -> Result<Self, Box<EvalAltResult>>

Macro generated variant of #fn_name that returns EvalAltResult. This allows the function to be used in scripts.

pub fn capture(self, c: String) -> Result<Self, QueryError>

Adds a capture to the last query. If no queries have been added, this will throw an error.

pub fn capture_script(self, c: String) -> Result<Self, Box<EvalAltResult>>

Macro generated variant of #fn_name that returns EvalAltResult. This allows the function to be used in scripts.

pub fn filter(self, f: FnPtr) -> Result<Self, QueryError>

Adds a capture to the last query. If no queries have been added, this will throw an error.

pub fn filter_script(self, f: FnPtr) -> Result<Self, Box<EvalAltResult>>

Macro generated variant of #fn_name that returns EvalAltResult. This allows the function to be used in scripts.

pub fn method_body_with_name(self, method_name: String) -> Self

Selects entire method body and returns

pub fn method_body_with_return_type(self, return_type: String) -> Self

Selects entire method body and returns

pub fn main_method(self) -> Self

Selects and returns the entire main method

pub fn class_body_with_name(self, class_name: String) -> Self

Selects entire class body with name

pub fn local_variables(self) -> Self

Selects local variable declaration statements

pub fn local_variables_with_name(self, name: String) -> Self

Selects local variable declaration statements with supplied name

pub fn local_variables_with_type(self, type_name: String) -> Self

Selects local variable declaration statements with supplied type

pub fn if_statements(self) -> Self

Selects if statements (entire, including else if and else)

pub fn for_loops(self) -> Self

Selects for loops

pub fn while_loops(self) -> Self

Selects while loops

pub fn method_invocations(self) -> Self

Selects method invocations

pub fn method_invocations_with_name(self, name: String) -> Self

Selects method invocations with supplied name

pub fn method_invocations_with_arguments(self, name: String) -> Self

Selects method invocations with supplied arguments

pub fn method_invocations_with_object(self, name: String) -> Self

Selects method invocations with supplied object

pub fn run_query(&self) -> Result<Dynamic, QueryError>

Runs the queries, and returns the result. TODO: Make it so that it doesn’t parse a new piece of code, just filters out the irrelevant line ranges. This performs better but more importantly is more accurate.

pub fn run_query_script(&self) -> Result<Dynamic, Box<EvalAltResult>>

Macro generated variant of #fn_name that returns EvalAltResult. This allows the function to be used in scripts.

pub fn grade_by_query(self) -> Result<GradeResult>

Grades the file according to the supplied queries, captures, and constraints.

pub fn grade_by_query_script(self) -> Result<GradeResult, Box<EvalAltResult>>

Macro generated variant of #fn_name that returns EvalAltResult. This allows the function to be used in scripts.