Struct umm::java::File

pub struct File {
    path: PathBuf,
    file_name: String,
    package_name: Option<String>,
    imports: Option<Vec<HashMap<String, String>>>,
    name: String,
    proper_name: String,
    test_methods: Vec<String>,
    kind: FileType,
    parser: Parser,
    description: String,
}

Struct representing a java file

Fields
path: PathBuf

path to java file.

file_name: String

name of file.

package_name: Option<String>

package the java file belongs to.

imports: Option<Vec<HashMap<String, String>>>

imports made by the java file.

name: String

name of the file TODO: How does this differ from file_name?

proper_name: String

colored terminal string representing java file name.

test_methods: Vec<String>

Name of tests methods in this file, as understood by JUnit.

kind: FileType

Name of tests methods in this file, colored using terminal color codes.

parser: Parser

The parser for this file

description: String

Conscise description of the file

Implementations

impl File

fn new(path: PathBuf) -> Result<Self>

Creates a new File from path

path: the path to read and try to create a File instance for.

pub fn new_script(path: PathBuf) -> Result<Self, Box<EvalAltResult>>

Macro generated variant of #fn_name that returns EvalAltResult. This allows the function to be used in scripts.

fn inner_doc_check(&self, err: Stdio, out: Stdio, in_: Stdio) -> Result<Output>

Returns the inner doc check of this File.

pub fn doc_check(&self) -> Result<String, JavaFileError>

Utility method to ask javac for documentation lints using the -Xdoclint flag.

The method simply returns the output produced by javac as a String. There is a ‘parse_diag method’ that can parse these to yield useful information.

pub fn doc_check_mut_script(&self) -> Result<String, Box<EvalAltResult>>

Utility method to ask javac for documentation lints using the -Xdoclint flag.

The method simply returns the output produced by javac as a String. There is a ‘parse_diag method’ that can parse these to yield useful information.

fn inner_check(&self, err: Stdio, out: Stdio, in_: Stdio) -> Result<Output>

Returns the inner check of this File.

pub fn check(&self) -> Result<String, JavaFileError>

Utility method to check for syntax errors using javac flag.

pub fn check_mut_script(&self) -> Result<String, Box<EvalAltResult>>

Utility method to check for syntax errors using javac flag.

fn inner_run( &self, input: Option<String>, err: Stdio, out: Stdio ) -> Result<Output>

Returns the inner run of this File.

pub fn run(&self, input: Option<String>) -> Result<String, JavaFileError>

Utility method to run a java file that has a main method.

pub fn run_mut_script( &self, input: Option<String> ) -> Result<String, Box<EvalAltResult>>

Utility method to run a java file that has a main method.

fn inner_test( &self, tests: Vec<&str>, err: Stdio, out: Stdio, in_: Stdio ) -> Result<Output>

Inner method to run tests.

pub fn test( &self, tests: Vec<&str>, project: Option<&Project> ) -> Result<String, JavaFileError>

A utility method that takes a list of strings (or types that implement Into<String>) meant to represent test method names, and runs those tests.

Returns the output from JUnit as a string. There are parsers in ‘parsers module’ that helps parse this output.

tests: list of strings (or types that implement Into<String>) meant to represent test method names,

pub fn test_mut_script( &mut self, tests: Vec<&str> ) -> Result<String, Box<EvalAltResult>>

A utility method that takes a list of strings (or types that implement Into<String>) meant to represent test method names, and runs those tests.

pub fn kind(&self) -> &FileType

A utility method that takes a list of strings (or types that implement Into<String>) meant to represent test method names, and runs those tests.

Returns the output from JUnit as a string. There are parsers in ‘parsers module’ that helps parse this output. Get a reference to the file’s kind.

pub fn file_name(&self) -> &str

Get a reference to the file’s file name.

pub fn test_methods(&self) -> Vec<String>

Get a reference to the file’s test methods.

pub fn test_methods_mut_script(&mut self) -> Array

Get a reference to the file’s test methods.

pub fn query(&self, q: &str) -> Result<Vec<HashMap<String, String>>>

treesitter query for this file

pub fn query_mut_script(&mut self, q: &str) -> Result<Array, Box<EvalAltResult>>

treesitter query for this file

pub fn path(&self) -> &PathBuf

Get a reference to the file’s path.

pub fn path_mut_script(&mut self) -> String

Get a reference to the file’s path.

pub fn package_name(&self) -> Option<&String>

Get a reference to the file’s proper name.

pub fn parser(&self) -> Parser

Get a reference to the file’s parser.

pub fn description(&self) -> String

Get a reference to the file’s description.

pub fn proper_name(&self) -> String

Get the file’s proper name.