Struct umm::vscode::Presentation

pub struct Presentation {
    reveal: Option<String>,
    reveal_problems: Option<String>,
    focus: Option<bool>,
    echo: Option<bool>,
    show_reuse_message: Option<bool>,
    panel: Option<String>,
    clear: Option<bool>,
    close: Option<bool>,
    group: Option<bool>,
}

Struct for VSCode task’s presentation.

Fields
reveal: Option<String>

Controls whether the Integrated Terminal panel is brought to front. Valid values are:

always - The panel is always brought to front. This is the default.

never - The user must explicitly bring the terminal panel to the front using the View > Terminal command (kb(workbench.action.terminal.toggleTerminal)).

silent - The terminal panel is brought to front only if the output is not scanned for errors and warnings.

reveal_problems: Option<String>

Controls whether the Problems panel is revealed when running this task or not. Takes precedence over option reveal. Default is never.

always - Always reveals the Problems panel when this task is executed.

onProblem - Only reveals the Problems panel if a problem is found.

never - Never reveals the Problems panel when this task is executed.

focus: Option<bool>

Controls whether the terminal is taking input focus or not. Default is false.

echo: Option<bool>

Controls whether the executed command is echoed in the terminal. Default is true.

show_reuse_message: Option<bool>

Controls whether to show the “Terminal will be reused by tasks, press any key to close it” message.

panel: Option<String>

Controls whether the terminal instance is shared between task runs. Possible values are:

shared - The terminal is shared and the output of other task runs are added to the same terminal.

dedicated - The terminal is dedicated to a specific task. If that task is executed again, the terminal is reused. However, the output of a different task is presented in a different terminal.

new - Every execution of that task is using a new clean terminal.

clear: Option<bool>

Controls whether the terminal is cleared before this task is run. Default is false.

close: Option<bool>

Controls whether the terminal the task runs in is closed when the task exits.

group: Option<bool>

Controls whether the task is executed in a specific terminal group using split panes. Tasks in the same group (specified by a string value) will use split terminals to present instead of a new terminal panel.

Implementations

impl Presentation

pub fn builder() -> PresentationBuilder<((), (), (), (), (), (), (), (), ())>

Create a builder for building Presentation. On the builder, call .reveal(...)(optional), .reveal_problems(...)(optional), .focus(...)(optional), .echo(...)(optional), .show_reuse_message(...)(optional), .panel(...)(optional), .clear(...)(optional), .close(...)(optional), .group(...)(optional) to set the values of the fields. Finally, call .build() to create the instance of Presentation.