Module umm::constants

A module defining a bunch of constant values to be used throughout

Structs

BUILD_DIR

Directory to store compiler artifacts

JAVA_TS_LANG

Reference to treesitter language struct

LIB_DIR

Directory for libraries, jars

POSTGREST_CLIENT

PostGrest client

ROOT_DIR

Path to project root

RUNTIME

Runtime

SEPARATOR

Platform specific separator character for javac paths

SOURCE_DIR

Directory for source files

SUPABASE_KEY

Supabase public api key

TEST_DIR

Directory for test files

UMM_DIR

Directory for umm artifacts

Constants

CLASSNAME_QUERY

Tree-sitter query that returns name of the class

COURSE

Current course. TODO: Move this to init script

IMPORT_QUERY

Tree-sitter query that returns imports made

INTERFACENAME_QUERY

Tree-sitter query that returns name of the interface

JUNIT_PLATFORM

file name for JUnit platform console standard jar

MAIN_METHOD_QUERY

Tree-sitter query to check the existence of a main method.

PACKAGE_QUERY

Tree-sitter query that returns name of the package

TERM

Current term. TODO: Move this to init script

TEST_ANNOTATION_QUERY

Tree-sitter query that returns name of the JUnit @Test annotated methods