Module umm::grade

For all things related to grading

Structs

ByHiddenTestGrader

Grades using hidden tests. Test file is downloaded, ran, and then cleaned up before returning.

ByUnitTestGrader

Grades by running tests, and reports how many tests pass. Final grade is the same percentage of maximum grade as the number of tests passing.

DocsGrader

A struct representing arguements to grade_docs function

Grade

A struct representing a grade

GradeResult

A struct to store grading results and display them

JavacDiagnostic

A struct representing a javac diagnostic message

JavacDiagnosticBuilder

Builder for JavacDiagnostic instances.

LineRef

A struct representing a line in a stack trace

MutationDiagnostic

A struct representing a PIT diagnostic message

MutationDiagnosticBuilder

Builder for MutationDiagnostic instances.

PromptRow

Schema for prompts table

UnitTestGrader

Runs mutation tests using Pitest to grade unit tests written by students.

__FieldVisitor 🔒

__FieldVisitor 🔒

__Visitor 🔒

__Visitor 🔒

Enums

__Field 🔒

__Field 🔒

Constants

FIELDS 🔒

FIELDS 🔒

Functions

generate_feedback

Generates a FEEDBACK file after prompting ChatGPT for feedback.

generate_feedback_script

Macro generated variant of #fn_name that returns EvalAltResult. This allows the function to be used in scripts.

get_source_context

Returns a ChatCompletionRequestMessage with the given line references that include contextual lines of code from the source

show_result

Print grade result